rename proj -> projection
