+++ stac_pydantic/extensions/projection.py
